Changed TopLevel to be an abstract class and moved it to abstract_classes. Updated all corresponding import statements.
