+++ core2/src/main/java/org/sbolstandard/core2/Model.java
@@ -3,6 +3,8 @@ package org.sbolstandard.core2;
 import java.net.URI;
 import java.util.Set;
 
+import org.sbolstandard.core2.abstract_classes.TopLevel;
+
 /**
  * @author Zhen Zhang
  * @author Nicholas Roehner
