+++ core2/src/main/java/org/sbolstandard/core2/abstract_classes/TopLevel.java
@@ -1,4 +1,4 @@
-package org.sbolstandard.core2;
+package org.sbolstandard.core2.abstract_classes;
 
 import java.net.URI;
